Update layout and add story state to top-level.
Introduces the StoryEditor and NoteStory to separate out the different
components. The top-level `text` is used by both the editor and the
story to render what the user is typing in the textarea.

Fixed regex to support punctuation after the NoteWord.

Added support for newlines in the text showing up as breaks in the
rendered document.

Preserved print layout and cleaned up styles.